+++ openapi_builder/converters/parameter/__init__.py
@@ -1,0 +1,3 @@
+"""Parameter converters are designed for processing the URL parameter."""
+
+PARAMETERS = []